+++ CHANGELOG.md
@@ -7,8 +7,9 @@
 - Support for testmode
   - Add tables for test entries
   - Handle Testmode to run executables
+  - Handle -b to record
 - Support for test macro
-  - Handle running all tests
+  - Handle running and recording all tests
 
 ## [0.0.9] - 2023-12-07